Print examples of chapter 1 of 'R for Dummies'.

To print a listing of all examples of a chapter, use ch1(). To run all the examples of ch1(), use example(ch1).

ch01()

ch1()

See also

Examples

# Chapter 1 - Introducing R: The Big Picture

# Recognizing the Benefits of Using R

## It comes as free, open-source code

### It runs anywhere

### It supports extensions

### It provides an engaged community

### It connects with other languages

# Looking At Some of the Unique Features of R

## Performing multiple calculations with vectors

x <- 1:5
x

#> [1] 1 2 3 4 5

x + 2

#> [1] 3 4 5 6 7

x + 6:10

#> [1]  7  9 11 13 15

## Processing more than just statistics

## Running code without a compiler