Print examples of chapter 3 of 'R for Dummies'.

To print a listing of all examples of a chapter, use ch3(). To run all the examples of ch3(), use example(ch3).

ch03()

ch3()

See also

Examples

# Chapter 3 - The Fundamentals of R

# Using the Full Power of Functions

## Vectorizing your functions

baskets.of.Granny <- c(12,4,4,6,9,3)
baskets.of.Granny
#> [1] 12  4  4  6  9  3
sum(baskets.of.Granny)
#> [1] 38

firstnames <- c("Joris", "Carolien", "Koen")
lastname <- "Meys"
paste(firstnames,lastname)
#> [1] "Joris Meys"    "Carolien Meys" "Koen Meys"    

authors <- c("Andrie","Joris")
lastnames <- c("de Vries","Meys")
paste(authors,lastnames)
#> [1] "Andrie de Vries" "Joris Meys"     

## Putting the argument in a function

# print() ### This line of code leads to deliberate error for illustration
print(x = "Isn't this fun?")
#> [1] "Isn't this fun?"

print(digits=4, x = 11/7)
#> [1] 1.571

# Making history

filename <- file.path(tempdir(), "Chapter3.Rhistory")
if (FALSE) {
savehistory(file = filename)
}
list.files(tempdir(), pattern = ".Rhistory")
#> character(0)
if (FALSE) {
loadhistory(file.path(tempdir(), "Chapter3.Rhistory"))
}

# Keeping Your Code Readable

## Following naming conventions

## Choosing a clear name

paste <- paste("This gets","confusing")
paste
#> [1] "This gets confusing"
paste("Don't","you","think?")
#> [1] "Don't you think?"

## Choosing a naming style

## Structuring your code

baskets.of.Geraldine <- c(5,3,2,2,12,9)
Intro <- "It is amazing! The All Star Grannies scored
a total of"

Outro <- "baskets in the last six games!"

Total.baskets <- baskets.of.Granny +
               baskets.of.Geraldine

Text <- paste(Intro,
              sum(Total.baskets),
              Outro)
cat(Text)
#> It is amazing! The All Star Grannies scored
#> a total of 71 baskets in the last six games!
Text
#> [1] "It is amazing! The All Star Grannies scored\na total of 71 baskets in the last six games!"

cat('If you doubt whether it works,
+ just try it out.')
#> If you doubt whether it works,
#> + just try it out.

## Adding comments

# The All Star Grannies do it again!
baskets.of.Granny <- c(12,4,4,6,9,3) # Granny rules
sum(baskets.of.Granny) # total number of points
#> [1] 38


# Getting from Base R to More

## Finding packages

## Installing packages

if (FALSE) {
install.packages("fortunes")
}

library("fortunes")
fortune("This is R")
#> 
#> Evelyn Hall: I would like to know how (if) I can extract some of the
#> information from the summary of my nlme.
#> Simon Blomberg: This is R. There is no if. Only how.
#>    -- Evelyn Hall and Simon 'Yoda' Blomberg
#>       R-help (April 2005)
#> 
fortune(161)
#> 
#> (3 times...rrrrgh...) and why do you think the mailing list is called
#> R-*packages* ???????????
#> Please do
#>   for(i in 1:20) cat("It's a package!\n")
#>    -- Martin Maechler (after a newly released *package* has been called
#>       *library* three times in its announcement on R-packages)
#>       R-help (March 2006)
#> 
detach(package:fortunes)